+++ src/config/local.js
@@ -1,0 +1,5 @@
+module.exports = {
+  db: {
+    host: '127.0.0.1',
+  },
+};