+++ src/lib/models-post-migration/apiToken.js
@@ -1,9 +1,19 @@
 module.exports = function (sequelize, DataTypes) {
   const Gw2ApiToken = sequelize.define('Gw2ApiToken', {
-    token: {
-      field: 'token',
-      type: DataTypes.STRING,
+    // REQUIRES MIGRATION (new field, new primary key)
+    id: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+      autoIncrement: true,
       primaryKey: true,
+    },
+
+    token: {
+      type: DataTypes.STRING,
+      // REQUIRES MIGRATION (primaryKey: false, allowNull false, unique true)
+      // primaryKey: true,
+      allowNull: false,
+      unique: true,
     },
     accountName: {
       type: DataTypes.STRING,
@@ -17,7 +27,6 @@ module.exports = function (sequelize, DataTypes) {
     },
     world: {
       type: DataTypes.INTEGER,
-      field: 'world',
       allowNull: false,
     },
     accountId: {
@@ -30,33 +39,24 @@ module.exports = function (sequelize, DataTypes) {
       type: DataTypes.BOOLEAN,
       allowNull: false,
       defaultValue: false,
-      field: 'primary',
     },
     guilds: {
       type: DataTypes.STRING,
       allowNull: true,
     },
-    // ----------- NEW FIELDS -----------
-    created: {
-      type: DataTypes.STRING,
-    },
-    access: {
-      type: DataTypes.STRING,
-    },
-    commander: {
+    created: DataTypes.STRING,
+    access: DataTypes.STRING,
+    commander: DataTypes.BOOLEAN,
+    fractalLevel: DataTypes.INTEGER,
+    dailyAp: DataTypes.INTEGER,
+    monthlyAp: DataTypes.INTEGER,
+    wvwRank: DataTypes.INTEGER,
+
+    // REQUIRES MIGRATION (new field)
+    stub: {
       type: DataTypes.BOOLEAN,
-    },
-    fractalLevel: {
-      type: DataTypes.INTEGER,
-    },
-    dailyAp: {
-      type: DataTypes.INTEGER,
-    },
-    monthlyAp: {
-      type: DataTypes.INTEGER,
-    },
-    wvwRank: {
-      type: DataTypes.INTEGER,
+      defaultValue: false,
+      allowNull: false,
     },
   }, {
     classMethods: {