+++ src/lib/models-post-migration/pvpStandings.js
@@ -1,20 +1,32 @@
 module.exports = (sequelize, DataTypes) => {
   return sequelize.define('PvpStandings', {
-    apiToken: {
-      type: DataTypes.STRING,
+    // REQUIRES MIGRATION (new field, new primary key)
+    id: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+      autoIncrement: true,
+      primaryKey: true,
+    },
+
+    // REQUIRE MIGRATION (apiToken -> apiTokenId
+    apiTokenId: {
+      // REQUIRED MIGRATION (STRING -> INT)
+      type: DataTypes.INTEGER,
       allowNull: false,
       onDelete: 'CASCADE',
-      primaryKey: true,
+      // REQUIRES MIGRATION (primaryKey: false)
+      // primaryKey: true,
       references: {
         model: 'Gw2ApiTokens',
-        key: 'token',
+        key: 'id',
       },
     },
 
     seasonId: {
       allowNull: false,
       type: DataTypes.STRING,
-      primaryKey: true,
+      // REQUIRES MIGRATION (primaryKey: false)
+      // primaryKey: true,
     },
 
     totalPointsCurrent: DataTypes.INTEGER,