+++ src/lib/models-post-migration/user.js
@@ -34,6 +34,13 @@ module.exports = function (sequelize, DataTypes) {
       allowNull: false,
       defaultValue: false,
     },
+
+    // REQUIRES MIGRATION
+    stub: {
+      type: DataTypes.BOOLEAN,
+      defaultValue: false,
+      allowNull: false,
+    },
   }, {
     classMethods: {
       associate (models) {